Права codestyle

+++ src/main/java/ru/yandex/practicum/filmorate/dal/mappers/ReviewMapper.java
@@ -11,11 +11,11 @@ import java.sql.SQLException;
 public class ReviewMapper implements RowMapper<Review> {
     @Override
     public Review mapRow(ResultSet rs, int rowNum) throws SQLException {
-        return new Review(rs.getInt("review_id"),
+        return new Review(rs.getLong("review_id"),
                 rs.getString("content"),
                 rs.getBoolean("is_positive"),
-                rs.getInt("user_id"),
-                rs.getInt("film_id"),
+                rs.getLong("user_id"),
+                rs.getLong("film_id"),
                 rs.getInt("useful"));
     }
 }
